[CTRACER]: Make class__remove_member remove bitfields as a group
Its not really being used right now, but for simplification, if we even ask
for bitfields to be removed, just do it in a group.

class__remove_member needs an overhaul, to handle things like bubbling up
bitfields when one of its entries is removed.

class_member__bitfield_tail seems to be solid and should be considered to be
used when moving bitfields around in dwarves_reorganize.c.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>