[CTRACER]: Add support for "pointer classes"
A pointer class is one that has as one of its members a pointer to the target class,
example:

target class: sock
pointer classes: sk_buff, socket

Because struct sk_buff and struct socket has members that are pointers to a struct sock.

This way we can follow all the struct socket and struct sk_buff methods and pass to
the struct sock collector as $skb->sk, for instance.

With this we now go all the way from the socket layer to the sk_buff layer, passing thru
the sock layer.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>