btf_loader: We can set class_member->type_offset earlier
Because we don't have to lookup the member type that is encoded after
it, i.e. some member fields depend on first processing everything, then
lookup the types to get the sizes, etc.

But the:

    member->byte_offset = member->bit_offset / 8;

Can be done soon after we have member->bit_offset loaded from the BTF
section.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>