btfdiff: Use --show_private_classes with DWARF
In DWARF we have the information if a struct/class is defined only
inside another struct/class or in a function, and then we consider those
to be 'private classes', requiring the use of --show_private_classes to
see those when asking for all structs.

That is not the case with BTF, that doesn't have that info and thus
shows all structs, private or not.

So, to compare the outputs in btfdiff we need to ask for
--show_private_classes when printing from DWARF.

With the Linux kernel vmlinux file the only private structure noticed
when not using this option, i.e. the only private class, as 'struct
sun_disklabel', defined in the block/partitions/sun.c file, inside the
'sun_partition' function.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>