dwarf_loader: Fixup bitfield entry with same number of bits as its base_type
For instance:

  $ cat examples/andrii/hsw_tsx_tuning.c
  // https://lore.kernel.org/bpf/CAEf4BzYoJgxYWX9iS7pda62cnGr5VOC70NGgqPurMNWP_w1daA@mail.gmail.com/

  typedef unsigned long long u64;
  typedef unsigned int u32;

  union hsw_tsx_tuning {
          struct {
                  u32 cycles_last_block     : 32,
                      hle_abort             : 1,
                      rtm_abort             : 1,
                      instruction_abort     : 1,
                      non_instruction_abort : 1,
                      retry                 : 1,
                      data_conflict         : 1,
                      capacity_writes       : 1,
                      capacity_reads        : 1;
          };
          u64        value;
  };

  int main() {
          union hsw_tsx_tuning t;
          return 0;
  }
  $

When recoding the DWARF bitfield sizes we look for a base type that has
32 bits and that has the same name, in this case:

  $ pahole --expand_types examples/andrii/hsw_tsx_tuning-gcc  | grep cycles_last_block
		/* typedef u32 */ unsigned int       cycles_last_block:32;               /*     0: 0  4 */
  $

We look for a 'unsigned int' with 32 bits, the same way as we do for all
the other bitfield entries. That worked well when the bit_size wasn't
the same as existing base type like 'unsigned int', which ended up
making that field get recoded to point to 'unsigned int', effectively
removing the original name, 'u32' in the case above, with the base type
it ultimately points to, 'unsigned int'.

Fix it making it use the same logic as for bit sizes that are different
from the base_type it points to, i.e. go ahead and create a new typedef
pointing to to a new base_type that has the right bit_size.

So, to sum up, before this fix:

$ btfdiff hsw_tsx_tuning-gcc
--- /tmp/btfdiff.dwarf.ErXffk   2019-02-25 10:26:56.863625697 -0300
+++ /tmp/btfdiff.btf.Y5EDdM     2019-02-25 10:26:56.864625706 -0300
@@ -1,6 +1,6 @@
 union hsw_tsx_tuning {
        struct {
-               unsigned int       cycles_last_block:32; /*     0: 0  4 */
+               u32                cycles_last_block:32; /*     0: 0  4 */
                u32                hle_abort:1;        /*     4:31  4 */
                u32                rtm_abort:1;        /*     4:30  4 */
                u32                instruction_abort:1; /*     4:29  4 */
$

Now btfdiff shows there are no differences, DWARF and BTF output produce
the same output, that matches the original struct.

The same tests were performed with vmlinux files produced by me and
Andrii.

Reported-by: Andrii Nakryiko <andriin@fb.com>
Reviewed-by: Andrii Nakryiko <andriin@fb.com>
Tested-by: Andrii Nakryiko <andriin@fb.com>
Cc: Alexei Starovoitov <ast@fb.com>
Cc: Yonghong Song <yhs@fb.com>
Cc: bpf@vger.kernel.org
Cc: dwarves@vger.kernel.org
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>